Create appointment API
