+++ examples/calendar/backend/services/appointments/routes.ts
@@ -6,6 +6,3 @@ export const appointmentsRouter = express.Router();
 
 /** POST /api/appointments */
 appointmentsRouter.route("/").post(controller.create);
-
-/** GET /api/appointments */
-//appointmentsRouter.route("/:userId").get(controller.get);
