+++ examples/calendar/backend/services/time/routes.ts
@@ -9,8 +9,5 @@ export const timeRouter = express.Router();
 /** POST /api/time */
 timeRouter.route("/").post(jwt(config), controller.create);
 
-/** PUT /api/time */
-timeRouter.route("/").put(jwt(config), controller.update);
-
 /** GET /api/time */
 timeRouter.route("/:userId").get(jwt(config), controller.get);
