+++ examples/calendar/backend/services/users/controller.ts
@@ -1,29 +1,9 @@
 import { NextFunction, Request, Response } from "express";
-
 import { User } from "../../models/user.model";
 
-export const find = (req: Request, res: Response, next: NextFunction) => {
-	// If a query string ?publicAddress=... is given, then filter results
-	const whereClause =
-		req.query && req.query.publicAddress
-			? {
-					where: { publicAddress: req.query.publicAddress }
-			  }
-			: undefined;
-
-	return User.findAll(whereClause)
-		.then((users: User[]) => res.json(users))
-		.catch(next);
-};
-
 export const get = (req: Request, res: Response, next: NextFunction) => {
-	// AccessToken payload is in req.user.payload, especially its `id` field
-	// UserId is the param in /users/:userId
-	// We only allow user accessing herself, i.e. require payload.id==userId
-	if ((req as any).user.payload.id !== +req.params.userId) {
-		return res.status(401).send({ error: "You can can only access yourself" });
-	}
-	return User.findByPk(req.params.userId)
+	
+	return User.findByPk(req.user.payload.id)
 		.then((user: User | null) => res.json(user))
 		.catch(next);
 };
@@ -33,26 +13,4 @@ export const create = (req: Request, res: Response, next: NextFunction) =>
 		.then((user: User) => res.json(user))
 		.catch(next);
 
-export const patch = (req: Request, res: Response, next: NextFunction) => {
-	// Only allow to fetch current user
-	if ((req as any).user.payload.id !== +req.params.userId) {
-		return res.status(401).send({ error: "You can can only access yourself" });
-	}
-	return User.findByPk(req.params.userId)
-		.then((user: User | null) => {
-			if (!user) {
-				return user;
-			}
 
-			Object.assign(user, req.body);
-			return user.save();
-		})
-		.then((user: User | null) => {
-			return user
-				? res.json(user)
-				: res.status(401).send({
-						error: `User with publicAddress ${req.params.userId} is not found in database`
-				  });
-		})
-		.catch(next);
-};
