+++ examples/calendar/backend/services/users/routes.ts
@@ -6,8 +6,6 @@ import * as controller from "./controller";
 
 export const userRouter = express.Router();
 
-/** GET /api/users */
-userRouter.route("/").get(controller.find);
 
 /** GET /api/users/:userId */
 /** Authenticated route */
@@ -16,6 +14,4 @@ userRouter.route("/:userId").get(jwt(config), controller.get);
 /** POST /api/users */
 userRouter.route("/").post(controller.create);
 
-/** PATCH /api/users/:userId */
-/** Authenticated route */
-userRouter.route("/:userId").patch(jwt(config), controller.patch);
+
